Close [WEI-37]

+++ app/src/main/java/com/leeweeder/weighttracker/domain/repository/DataStoreRepository.kt
@@ -3,8 +3,8 @@ package com.leeweeder.weighttracker.domain.repository
 import kotlinx.coroutines.flow.Flow
 
 interface DataStoreRepository {
-    suspend fun saveGoalWeight(value: Float)
-    fun readGoalWeightState(): Flow<Float>
+    suspend fun saveGoalWeight(value: Int)
+    fun readGoalWeightState(): Flow<Int>
 
     suspend fun saveShouldHideOnBoarding(shouldHideOnBoarding: Boolean)
     fun readOnBoardingState(): Flow<Boolean>
